requirements/base.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.1
django-admin-tools==0.6.0
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.8.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
-e git://github.com/smacker/django-filebrowser-no-grappelli.git@25613ff879325f2da8e465c028b9657ae6dd79cf#egg=django_filebrowser_no_grappelli-master
django-filter==0.9.2
django-grappelli==2.7.1
django-multiupload==0.3
djangorestframework==3.2.0
kombu==3.0.26
Markdown==2.6.2
Pillow==2.8.1
psycopg2==2.6
pytz==2015.4
six==1.9.0
sqlparse==0.1.15